Allow all Scala 3 tests to compile and pass
Previously when exploring an ADT tree all case object would
be resolved as `object A` symbols which turned out to be a term, which
did not play well with the implicit resolution. Now they were changed to
`module class A$` type symbols. Additionally referencing case objects
was corrected, as the previous way would result in "not a term"
assertion errors.